prettified avalanche resolver

+++ contracts/resolver/avalanche/helpers.sol
@@ -5,23 +5,32 @@ import {Variables} from "./variables.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract Helper is Variables {
-    function getAaveAvailability(address[] memory  _tokens, uint256[] memory  _amounts) internal view returns (bool) {
-        for(uint256 i = 0; i < _tokens.length; i++) {
+    function getAaveAvailability(
+        address[] memory _tokens,
+        uint256[] memory _amounts
+    ) internal view returns (bool) {
+        for (uint256 i = 0; i < _tokens.length; i++) {
             IERC20 token_ = IERC20(_tokens[i]);
-            (,,,,,,,,bool isActive,) = aaveProtocolDataProvider.getReserveConfigurationData(_tokens[i]);
-            (address aTokenAddr,,) = aaveProtocolDataProvider.getReserveTokensAddresses(_tokens[i]);
-            if(isActive == false) return false;
-            if(token_.balanceOf(aTokenAddr) < _amounts[i]) return false;
+            (, , , , , , , , bool isActive, ) = aaveProtocolDataProvider
+                .getReserveConfigurationData(_tokens[i]);
+            (address aTokenAddr, , ) = aaveProtocolDataProvider
+                .getReserveTokensAddresses(_tokens[i]);
+            if (isActive == false) return false;
+            if (token_.balanceOf(aTokenAddr) < _amounts[i]) return false;
         }
         return true;
     }
 
-    function getRoutesWithAvailability(uint16[] memory _routes, address[] memory _tokens, uint256[] memory _amounts) internal view returns (uint16[] memory) {
+    function getRoutesWithAvailability(
+        uint16[] memory _routes,
+        address[] memory _tokens,
+        uint256[] memory _amounts
+    ) internal view returns (uint16[] memory) {
         uint16[] memory routesWithAvailability_ = new uint16[](7);
-        uint j = 0;
-        for(uint256 i = 0; i < _routes.length; i++) {
+        uint256 j = 0;
+        for (uint256 i = 0; i < _routes.length; i++) {
             if (_routes[i] == 1) {
-                if(getAaveAvailability(_tokens, _amounts)) {
+                if (getAaveAvailability(_tokens, _amounts)) {
                     routesWithAvailability_[j] = _routes[i];
                     j++;
                 }
@@ -32,11 +41,25 @@ contract Helper is Variables {
         return routesWithAvailability_;
     }
 
-    function bubbleSort(address[] memory _tokens, uint256[] memory _amounts) internal pure returns (address[] memory, uint256[] memory) {
+    function bubbleSort(address[] memory _tokens, uint256[] memory _amounts)
+        internal
+        pure
+        returns (address[] memory, uint256[] memory)
+    {
         for (uint256 i = 0; i < _tokens.length - 1; i++) {
-            for( uint256 j = 0; j < _tokens.length - i - 1 ; j++) {
-                if(_tokens[j] > _tokens[j+1]) {
-                    (_tokens[j], _tokens[j+1], _amounts[j], _amounts[j+1]) = (_tokens[j+1], _tokens[j], _amounts[j+1], _amounts[j]);
+            for (uint256 j = 0; j < _tokens.length - i - 1; j++) {
+                if (_tokens[j] > _tokens[j + 1]) {
+                    (
+                        _tokens[j],
+                        _tokens[j + 1],
+                        _amounts[j],
+                        _amounts[j + 1]
+                    ) = (
+                        _tokens[j + 1],
+                        _tokens[j],
+                        _amounts[j + 1],
+                        _amounts[j]
+                    );
                 }
             }
         }
@@ -44,8 +67,8 @@ contract Helper is Variables {
     }
 
     function validateTokens(address[] memory _tokens) internal pure {
-        for (uint i = 0; i < _tokens.length - 1; i++) {
-            require(_tokens[i] != _tokens[i+1], "non-unique-tokens");
+        for (uint256 i = 0; i < _tokens.length - 1; i++) {
+            require(_tokens[i] != _tokens[i + 1], "non-unique-tokens");
         }
     }
 }
